Implemented ability to delete relationship between user and facility without delete user action

+++ app/Http/Controllers/UserController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\API\UserDeleteRequest;
 use App\Http\Requests\API\UserUpdateRequest;
+use App\Models\FacilityUser;
 use App\Models\User;
 use App\Repositories\UserRepository;
 use App\Services\UserService;
@@ -52,8 +54,9 @@ class UserController extends Controller
         return response()->json(config('app.user_roles'));
     }
 
-    public function destroy(int $id): JsonResponse
+    public function destroy(int $id, UserDeleteRequest $request): JsonResponse
     {
+        $request = $request->validated();
         $user = User::find($id);
         if (!$user) {
             return response()->json(['User not found'], 400);
@@ -63,7 +66,18 @@ class UserController extends Controller
             return response()->json(['message' => 'Permission denied'], 403);
         }
 
-        $user->delete();
+        if (isset($request['facility_id'])) {
+            if (!FacilityUser::where('facility_id', $request['facility_id'])->where('user_id', $id)->first()) {
+                return response()->json(['message' => 'The user with ID ' . $id . ' is not assigned to the facility with ID ' . $request['facility_id']], 400);
+            }
+            if (FacilityUser::where('user_id', $id)->count() === 1) {
+                $user->delete();
+            } else {
+                FacilityUser::where('facility_id', $request['facility_id'])->where('user_id', $id)->delete();
+            }
+        } else {
+            $user->delete();
+        }
         return response()->json('1', 204);
     }
 }
